Split staff mutations
