+++ saleor/graphql/account/mutations/account/__init__.py
@@ -6,6 +6,7 @@ from .account_register import AccountRegister
 from .account_request_deletion import AccountRequestDeletion
 from .account_set_default_address import AccountSetDefaultAddress
 from .account_update import AccountUpdate
+from .confirm_account import ConfirmAccount
 from .confirm_email_change import ConfirmEmailChange
 from .request_email_change import RequestEmailChange
 
@@ -18,6 +19,7 @@ __all__ = [
     "AccountRequestDeletion",
     "AccountSetDefaultAddress",
     "AccountUpdate",
+    "ConfirmAccount",
     "ConfirmEmailChange",
     "RequestEmailChange",
 ]
