+++ saleor/graphql/account/mutations/authentication/__init__.py
@@ -5,7 +5,10 @@ from .external_logout import ExternalLogout
 from .external_obtain_access_tokens import ExternalObtainAccessTokens
 from .external_refresh import ExternalRefresh
 from .external_verify import ExternalVerify
+from .password_change import PasswordChange
 from .refresh_token import RefreshToken
+from .request_password_reset import RequestPasswordReset
+from .set_password import SetPassword
 from .verify_token import VerifyToken
 
 __all__ = [
@@ -16,6 +19,9 @@ __all__ = [
     "ExternalObtainAccessTokens",
     "ExternalRefresh",
     "ExternalVerify",
+    "PasswordChange",
     "RefreshToken",
+    "RequestPasswordReset",
+    "SetPassword",
     "VerifyToken",
 ]
