+++ saleor/graphql/account/mutations/base.py
@@ -1,50 +1,36 @@
-from typing import cast
+from collections import defaultdict
+from typing import List
 
 import graphene
-from django.conf import settings
-from django.contrib.auth import password_validation
-from django.contrib.auth.hashers import make_password
-from django.contrib.auth.tokens import default_token_generator
-from django.core.exceptions import ObjectDoesNotExist, ValidationError
-from django.utils import timezone
+from django.core.exceptions import ValidationError
 
 from ....account import events as account_events
-from ....account import models
 from ....account.error_codes import AccountErrorCode
-from ....account.notifications import (
-    send_password_reset_notification,
-    send_set_password_notification,
-)
+from ....account.notifications import send_set_password_notification
 from ....account.search import prepare_user_search_document_value
-from ....account.utils import retrieve_user_by_email
 from ....checkout import AddressType
 from ....core.exceptions import PermissionDenied
 from ....core.tracing import traced_atomic_transaction
 from ....core.utils.url import validate_storefront_url
-from ....giftcard.utils import assign_user_gift_cards
 from ....graphql.utils import get_user_or_app_from_context
-from ....order.utils import match_orders_with_new_user
 from ....permission.auth_filters import AuthorizationFilters
 from ....permission.enums import AccountPermissions
 from ...account.i18n import I18nMixin
 from ...account.types import Address, AddressInput, User
 from ...app.dataloaders import get_app_promise
 from ...channel.utils import clean_channel, validate_channel
-from ...core import ResolveInfo
-from ...core.context import disallow_replica_in_context
+from ...core import ResolveInfo, SaleorContext
 from ...core.descriptions import ADDED_IN_310, ADDED_IN_314
 from ...core.doc_category import DOC_CATEGORY_USERS
 from ...core.enums import LanguageCodeEnum
-from ...core.mutations import (
-    BaseMutation,
-    ModelDeleteMutation,
-    ModelMutation,
-    validation_error_to_error_type,
-)
-from ...core.types import AccountError, BaseInputObjectType, NonNullList
+from ...core.mutations import ModelDeleteMutation, ModelMutation
+from ...core.types import BaseInputObjectType, NonNullList
 from ...meta.inputs import MetadataInput
 from ...plugins.dataloaders import get_plugin_manager_promise
-from .authentication import CreateToken
+from ..utils import (
+    get_not_manageable_permissions_when_deactivate_or_remove_users,
+    get_out_of_scope_users,
+)
 
 BILLING_ADDRESS_FIELD = "default_billing_address"
 SHIPPING_ADDRESS_FIELD = "default_shipping_address"
@@ -70,288 +56,6 @@ def check_can_edit_address(context, address):
     raise PermissionDenied(
         permissions=[AccountPermissions.MANAGE_USERS, AuthorizationFilters.OWNER]
     )
-
-
-class AccountBaseInput(BaseInputObjectType):
-    first_name = graphene.String(description="Given name.")
-    last_name = graphene.String(description="Family name.")
-    language_code = graphene.Argument(
-        LanguageCodeEnum, required=False, description="User language code."
-    )
-
-
-class SetPassword(CreateToken):
-    class Arguments:
-        token = graphene.String(
-            description="A one-time token required to set the password.", required=True
-        )
-        email = graphene.String(required=True, description="Email of a user.")
-        password = graphene.String(required=True, description="Password of a user.")
-
-    class Meta:
-        description = (
-            "Sets the user's password from the token sent by email "
-            "using the RequestPasswordReset mutation."
-        )
-        doc_category = DOC_CATEGORY_USERS
-        error_type_class = AccountError
-        error_type_field = "account_errors"
-
-    @classmethod
-    def mutate(  # type: ignore[override]
-        cls, root, info: ResolveInfo, /, *, email, password, token
-    ):
-        disallow_replica_in_context(info.context)
-        manager = get_plugin_manager_promise(info.context).get()
-        result = manager.perform_mutation(
-            mutation_cls=cls,
-            root=root,
-            info=info,
-            data={"email": email, "password": password, "token": token},
-        )
-        if result is not None:
-            return result
-
-        try:
-            cls._set_password_for_user(email, password, token)
-        except ValidationError as e:
-            errors = validation_error_to_error_type(e, AccountError)
-            return cls.handle_typed_errors(errors)
-        return super().mutate(root, info, email=email, password=password)
-
-    @classmethod
-    def _set_password_for_user(cls, email, password, token):
-        try:
-            user = models.User.objects.get(email=email)
-        except ObjectDoesNotExist:
-            raise ValidationError(
-                {
-                    "email": ValidationError(
-                        "User doesn't exist", code=AccountErrorCode.NOT_FOUND.value
-                    )
-                }
-            )
-        if not default_token_generator.check_token(user, token):
-            raise ValidationError(
-                {
-                    "token": ValidationError(
-                        INVALID_TOKEN, code=AccountErrorCode.INVALID.value
-                    )
-                }
-            )
-        try:
-            password_validation.validate_password(password, user)
-        except ValidationError as error:
-            raise ValidationError({"password": error})
-        user.set_password(password)
-        user.save(update_fields=["password", "updated_at"])
-        account_events.customer_password_reset_event(user=user)
-
-
-class RequestPasswordReset(BaseMutation):
-    class Arguments:
-        email = graphene.String(
-            required=True,
-            description="Email of the user that will be used for password recovery.",
-        )
-        redirect_url = graphene.String(
-            required=True,
-            description=(
-                "URL of a view where users should be redirected to "
-                "reset the password. URL in RFC 1808 format."
-            ),
-        )
-        channel = graphene.String(
-            description=(
-                "Slug of a channel which will be used for notify user. Optional when "
-                "only one channel exists."
-            )
-        )
-
-    class Meta:
-        description = "Sends an email with the account password modification link."
-        doc_category = DOC_CATEGORY_USERS
-        error_type_class = AccountError
-        error_type_field = "account_errors"
-
-    @classmethod
-    def clean_user(cls, email, redirect_url):
-        try:
-            validate_storefront_url(redirect_url)
-        except ValidationError as error:
-            raise ValidationError(
-                {"redirect_url": error}, code=AccountErrorCode.INVALID.value
-            )
-
-        user = retrieve_user_by_email(email)
-        if not user:
-            raise ValidationError(
-                {
-                    "email": ValidationError(
-                        "User with this email doesn't exist",
-                        code=AccountErrorCode.NOT_FOUND.value,
-                    )
-                }
-            )
-
-        if not user.is_active:
-            raise ValidationError(
-                {
-                    "email": ValidationError(
-                        "User with this email is inactive",
-                        code=AccountErrorCode.INACTIVE.value,
-                    )
-                }
-            )
-
-        if password_reset_time := user.last_password_reset_request:
-            delta = timezone.now() - password_reset_time
-            if delta.total_seconds() < settings.RESET_PASSWORD_LOCK_TIME:
-                raise ValidationError(
-                    {
-                        "email": ValidationError(
-                            "Password reset already requested",
-                            code=AccountErrorCode.PASSWORD_RESET_ALREADY_REQUESTED.value,
-                        )
-                    }
-                )
-
-        return user
-
-    @classmethod
-    def perform_mutation(cls, _root, info: ResolveInfo, /, **data):
-        email = data["email"]
-        redirect_url = data["redirect_url"]
-        channel_slug = data.get("channel")
-        user = cls.clean_user(email, redirect_url)
-
-        if not user.is_staff:
-            channel_slug = clean_channel(
-                channel_slug, error_class=AccountErrorCode
-            ).slug
-        elif channel_slug is not None:
-            channel_slug = validate_channel(
-                channel_slug, error_class=AccountErrorCode
-            ).slug
-        manager = get_plugin_manager_promise(info.context).get()
-        send_password_reset_notification(
-            redirect_url,
-            user,
-            manager,
-            channel_slug=channel_slug,
-            staff=user.is_staff,
-        )
-        user.last_password_reset_request = timezone.now()
-        user.save(update_fields=["last_password_reset_request"])
-        return RequestPasswordReset()
-
-
-class ConfirmAccount(BaseMutation):
-    user = graphene.Field(User, description="An activated user account.")
-
-    class Arguments:
-        token = graphene.String(
-            description="A one-time token required to confirm the account.",
-            required=True,
-        )
-        email = graphene.String(
-            description="E-mail of the user performing account confirmation.",
-            required=True,
-        )
-
-    class Meta:
-        description = (
-            "Confirm user account with token sent by email during registration."
-        )
-        doc_category = DOC_CATEGORY_USERS
-        error_type_class = AccountError
-        error_type_field = "account_errors"
-
-    @classmethod
-    def perform_mutation(cls, _root, info: ResolveInfo, /, **data):
-        try:
-            user = models.User.objects.get(email=data["email"])
-        except ObjectDoesNotExist:
-            raise ValidationError(
-                {
-                    "email": ValidationError(
-                        "User with this email doesn't exist",
-                        code=AccountErrorCode.NOT_FOUND.value,
-                    )
-                }
-            )
-
-        if not default_token_generator.check_token(user, data["token"]):
-            raise ValidationError(
-                {
-                    "token": ValidationError(
-                        INVALID_TOKEN, code=AccountErrorCode.INVALID.value
-                    )
-                }
-            )
-
-        user.is_active = True
-        user.save(update_fields=["is_active", "updated_at"])
-
-        match_orders_with_new_user(user)
-        assign_user_gift_cards(user)
-
-        return ConfirmAccount(user=user)
-
-
-class PasswordChange(BaseMutation):
-    user = graphene.Field(User, description="A user instance with a new password.")
-
-    class Arguments:
-        old_password = graphene.String(
-            required=False, description="Current user password."
-        )
-        new_password = graphene.String(required=True, description="New user password.")
-
-    class Meta:
-        description = "Change the password of the logged in user."
-        doc_category = DOC_CATEGORY_USERS
-        error_type_class = AccountError
-        error_type_field = "account_errors"
-        permissions = (AuthorizationFilters.AUTHENTICATED_USER,)
-
-    @staticmethod
-    def raise_invalid_credentials():
-        raise ValidationError(
-            {
-                "old_password": ValidationError(
-                    "Old password isn't valid.",
-                    code=AccountErrorCode.INVALID_CREDENTIALS.value,
-                )
-            }
-        )
-
-    @classmethod
-    def perform_mutation(cls, _root, info: ResolveInfo, /, **data):
-        user = info.context.user
-        user = cast(models.User, user)
-        old_password = data.get("old_password")
-        new_password = data["new_password"]
-
-        if old_password is None:
-            # Spend time hashing useless password
-            # This prevents the outside actors from telling if user has
-            # unusable password set or not by measuring API's response time
-            make_password("waste-time")
-
-            if user.has_usable_password():
-                cls.raise_invalid_credentials()
-        elif not user.check_password(old_password):
-            cls.raise_invalid_credentials()
-        try:
-            password_validation.validate_password(new_password, user)
-        except ValidationError as error:
-            raise ValidationError({"new_password": error})
-
-        user.set_password(new_password)
-        user.save(update_fields=["password", "updated_at"])
-        account_events.customer_password_changed_event(user=user)
-        return PasswordChange(user=user)
 
 
 class BaseAddressUpdate(ModelMutation, I18nMixin):
@@ -633,3 +337,145 @@ class BaseCustomerCreate(ModelMutation, I18nMixin):
         if cleaned_input.get("metadata"):
             manager = get_plugin_manager_promise(info.context).get()
             cls.call_event(manager.customer_metadata_updated, instance)
+
+
+class UserDeleteMixin:
+    class Meta:
+        abstract = True
+
+    @classmethod
+    def clean_instance(cls, info: ResolveInfo, instance) -> None:
+        user = info.context.user
+        if instance == user:
+            raise ValidationError(
+                {
+                    "id": ValidationError(
+                        "You cannot delete your own account.",
+                        code=AccountErrorCode.DELETE_OWN_ACCOUNT.value,
+                    )
+                }
+            )
+        elif instance.is_superuser:
+            raise ValidationError(
+                {
+                    "id": ValidationError(
+                        "Cannot delete this account.",
+                        code=AccountErrorCode.DELETE_SUPERUSER_ACCOUNT.value,
+                    )
+                }
+            )
+
+
+class CustomerDeleteMixin(UserDeleteMixin):
+    class Meta:
+        abstract = True
+
+    @classmethod
+    def clean_instance(cls, info: ResolveInfo, instance) -> None:
+        super().clean_instance(info, instance)
+        if instance.is_staff:
+            raise ValidationError(
+                {
+                    "id": ValidationError(
+                        "Cannot delete a staff account.",
+                        code=AccountErrorCode.DELETE_STAFF_ACCOUNT.value,
+                    )
+                }
+            )
+
+    @classmethod
+    def post_process(cls, info: ResolveInfo, deleted_count=1):
+        app = get_app_promise(info.context).get()
+        account_events.customer_deleted_event(
+            staff_user=info.context.user,
+            app=app,
+            deleted_count=deleted_count,
+        )
+
+
+class StaffDeleteMixin(UserDeleteMixin):
+    class Meta:
+        abstract = True
+
+    @classmethod
+    def check_permissions(cls, context: SaleorContext, permissions=None, **data):
+        if get_app_promise(context).get():
+            raise PermissionDenied(
+                message="Apps are not allowed to perform this mutation."
+            )
+        return super().check_permissions(context, permissions)  # type: ignore[misc] # mixin # noqa: E501
+
+    @classmethod
+    def clean_instance(cls, info: ResolveInfo, instance):
+        errors: defaultdict[str, List[ValidationError]] = defaultdict(list)
+
+        requestor = info.context.user
+
+        cls.check_if_users_can_be_deleted(info, [instance], "id", errors)
+        cls.check_if_requestor_can_manage_users(requestor, [instance], "id", errors)
+        cls.check_if_removing_left_not_manageable_permissions(
+            requestor, [instance], "id", errors
+        )
+        if errors:
+            raise ValidationError(errors)
+
+    @classmethod
+    def check_if_users_can_be_deleted(cls, info: ResolveInfo, instances, field, errors):
+        """Check if only staff users will be deleted. Cannot delete non-staff users."""
+        not_staff_users = set()
+        for user in instances:
+            if not user.is_staff:
+                not_staff_users.add(user)
+            try:
+                super().clean_instance(info, user)
+            except ValidationError as error:
+                errors["ids"].append(error)
+
+        if not_staff_users:
+            user_pks = [
+                graphene.Node.to_global_id("User", user.pk) for user in not_staff_users
+            ]
+            msg = "Cannot delete a non-staff users."
+            code = AccountErrorCode.DELETE_NON_STAFF_USER.value
+            params = {"users": user_pks}
+            errors[field].append(ValidationError(msg, code=code, params=params))
+
+    @classmethod
+    def check_if_requestor_can_manage_users(cls, requestor, instances, field, errors):
+        """Requestor can't manage users with wider scope of permissions."""
+        if requestor.is_superuser:
+            return
+        out_of_scope_users = get_out_of_scope_users(requestor, instances)
+        if out_of_scope_users:
+            user_pks = [
+                graphene.Node.to_global_id("User", user.pk)
+                for user in out_of_scope_users
+            ]
+            msg = "You can't manage this users."
+            code = AccountErrorCode.OUT_OF_SCOPE_USER.value
+            params = {"users": user_pks}
+            error = ValidationError(msg, code=code, params=params)
+            errors[field] = error
+
+    @classmethod
+    def check_if_removing_left_not_manageable_permissions(
+        cls, requestor, users, field, errors: defaultdict[str, List[ValidationError]]
+    ):
+        """Check if after removing users all permissions will be manageable.
+
+        After removing users, for each permission, there should be at least one
+        active staff member who can manage it (has both “manage staff” and
+        this permission).
+        """
+        if requestor.is_superuser:
+            return
+        permissions = get_not_manageable_permissions_when_deactivate_or_remove_users(
+            users
+        )
+        if permissions:
+            # add error
+            msg = "Users cannot be removed, some of permissions will not be manageable."
+            code = AccountErrorCode.LEFT_NOT_MANAGEABLE_PERMISSION.value
+            params = {"permissions": permissions}
+            error = ValidationError(msg, code=code, params=params)
+            errors[field] = [error]
