+++ saleor/graphql/account/mutations/staff/base.py
@@ -1,0 +1,7 @@
+from ....core.mutations import ModelDeleteMutation, ModelWithExtRefMutation
+from ..base import UserDeleteMixin
+
+
+class UserDelete(UserDeleteMixin, ModelDeleteMutation, ModelWithExtRefMutation):
+    class Meta:
+        abstract = True
